improvement: ESPNOW sending rpm

+++ include/esp_now_tool.h
@@ -49,7 +49,7 @@ typedef struct {
     uint8_t state;                        //Indicate that if has received broadcast ESPNOW data or not.
     uint16_t crc;                         //CRC16 value of ESPNOW data.
     uint32_t magic;                       //Magic number which is used to determine which device to send unicast ESPNOW data.
-    uint8_t payload[0];                   //Real payload of ESPNOW data.
+    uint16_t payload[0];                   //Real payload of ESPNOW data.
 } __attribute__((packed)) esp_now_data_t;
 
 /* Parameters of sending ESPNOW data. */
